MAINT: remove deprecation warning for **kwargs
